+++ src/main/java/types/Meeting.java
@@ -7,7 +7,7 @@ public class Meeting {
   @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss")
   public Timestamp beginDate; // contains date and time of first event.
 
-  public long minutesDuration; // Duration of meeting
+  public int minutesDuration; // Duration of meeting
 
   @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss")
   public Timestamp endDate; // When the meeting stops repeating
